Use 5ms as a more modern value for weightestimate
The old UNIX box I wrote this on took 20ms to transfer and render a
frame of sixels. My current box takes about 4ms, which means it takes
quite a while (several seconds) before the weight estimate is updated
and the animation plays at the right speed.

A better algorithm for maintaining the frame rate is called for, but
in the meantime, this "fixes" it for me.