Disable -icanon to fix ^? when in benchmark mode
In benchmark mode, the 'read' command waits only 1ms for a key.

That means keys that keys will be hit when 'read' is not running and
have to be buffered. However, when read isn't running the tty was in
icanon mode, so it performed line editing actions, interpreting and
consuming the backspace key.