Use "nanosecond precision" timing for overall FPS
Previously, we used seconds, which meant FPS was not very accurate for
short test runs.

Of course, it's not nanosecond *accuracy*, even if we have precision
down to nine decimal places. We only use nanoseconds since they are
the easiest time stamp to get which is a fraction of a second.

This change makes the average frames per second that is shown at the
end when benchmarking converge on the correct answer very quickly.