Add -E flag to exceed 1000px limit of XTerm
By default the -e, --exact flag is not actually exact.

It still limits the output to 1000px by 1000px because XTerm(369)
shows a blank screen instead of truncating the image for images larger
than that.

The new -E flag disables that. I expect this will eventually be moot
once XTerm fixes its bug.