add --stupid-fast to go faster than possible
Sends frames without waiting for the terminal to have finished
processing the previous frame.

Note that "stupid" can be used in a positive sense in common American
English slang as way to express that something is extreme.
For instance,

    "My new car goes so fast, it's stupid!"

The option --stupid-fast plays on both meanings because it is both
extremely fast and it is also not a very good idea.

Besides that FPS will be wrong, --stupid-fast packs the buffer to the
terminal with so many frames that it is difficult to quit. (I just
tried it over an ssh link, and even Ctrl+C was not able to halt it
immediately. That may be a bug in ssh's flusho implementation.)